move AppGroupTraverser to core project for more access
